Moved draw API into its own module
