+++ src/gui/backend_glium.rs
@@ -141,8 +141,8 @@ impl super::Drawable for Primitive {
             .magnify_filter(glium::uniforms::MagnifySamplerFilter::Nearest);
 
         let uniforms = glium::uniform! {
-            world_transform: dcf.state.world_transform.to_cols_array_2d(),
-            color_multiplier_global: dcf.state.color_multiplier.0.to_array(),
+            world_transform: dcf.state().world_transform.to_cols_array_2d(),
+            color_multiplier_global: dcf.state().color_multiplier.0.to_array(),
             tex: sampler,
         };
